Widen sidebar by the rail width so content keeps its full width
The rail column was starving sidebar content (group labels truncated).
AppLayout now offsets --sidebar-width by SERVER_RAIL_WIDTH_PX while the rail
is visible, including during resize drags; browser-tab views resync bounds
when the rail toggles.
