+++ apps/app/src/components/layout/AppLayout.tsx
@@ -54,6 +54,8 @@ import {
   shouldUseMacosDesktopChrome,
 } from "@/lib/bb-desktop";
 import { useDesktopWindowState } from "@/hooks/useDesktopWindowState";
+import { useHasServerRail } from "@/hooks/useDesktopServerList";
+import { SERVER_RAIL_WIDTH_PX } from "@/components/sidebar/ServerRail";
 import {
   getLegacyProjectComposeRoutePath,
   getProjectArchivedRoutePath,
@@ -495,6 +497,8 @@ export function AppLayout({ children }: AppLayoutProps) {
   const startXRef = useRef(0);
   const startWidthRef = useRef(0);
   const liveWidthRef = useRef(sidebarWidth);
+  const railWidthOffsetRef = useRef(0);
+  const hasServerRail = useHasServerRail();
   const animationFrameRef = useRef<number | null>(null);
   const showHeader = !isThreadView && !isRootView;
   const [desktopInfo] = useState(getBbDesktopInfo);
@@ -504,8 +508,14 @@ export function AppLayout({ children }: AppLayoutProps) {
     desktopInfo,
     windowState: desktopWindowState,
   });
+  // The server rail lives inside the sidebar panel; widen the panel by the
+  // rail width so the sidebar content keeps its full persisted width. The
+  // resize handlers below write the CSS var directly during drags, so they
+  // apply the same offset via railWidthOffsetRef.
+  const railWidthOffset = hasServerRail ? SERVER_RAIL_WIDTH_PX : 0;
+  railWidthOffsetRef.current = railWidthOffset;
   const sidebarProviderStyle: SidebarProviderStyle = {
-    "--sidebar-width": `${sidebarWidth}px`,
+    "--sidebar-width": `${sidebarWidth + railWidthOffset}px`,
   };
 
   const project = projectId
@@ -641,7 +651,7 @@ export function AppLayout({ children }: AppLayoutProps) {
     }
     providerRef.current?.style.setProperty(
       "--sidebar-width",
-      `${liveWidthRef.current}px`,
+      `${liveWidthRef.current + railWidthOffsetRef.current}px`,
     );
     dispatchBrowserViewBoundsSync();
     setSidebarWidth(liveWidthRef.current);
@@ -656,7 +666,7 @@ export function AppLayout({ children }: AppLayoutProps) {
       animationFrameRef.current = null;
       providerRef.current?.style.setProperty(
         "--sidebar-width",
-        `${liveWidthRef.current}px`,
+        `${liveWidthRef.current + railWidthOffsetRef.current}px`,
       );
       dispatchBrowserViewBoundsSync();
     };
@@ -696,6 +706,12 @@ export function AppLayout({ children }: AppLayoutProps) {
   useEffect(() => {
     liveWidthRef.current = sidebarWidth;
   }, [sidebarWidth]);
+
+  useEffect(() => {
+    // Rail appearing/disappearing changes the effective sidebar width; keep
+    // native browser-tab views in sync just like a sidebar toggle does.
+    window.requestAnimationFrame(dispatchBrowserViewBoundsSync);
+  }, [hasServerRail]);
 
   useEffect(() => {
     if (typeof document === "undefined") return;
